requirements.txt
flask
firebase-admin
openai
python-dotenv
google-generativeai
google-genai
